requirements.txt
torchvision==0.7.0
torch==1.6.0
tensorboard-logger
tqdm
Pillow
scikit-image
scikit-learn
scipy